# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/access-token": {
            "get": {
                "description": "列出访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "列出访问令牌",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页起始位置",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页每页数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "列出成功",
                        "schema": {
                            "$ref": "#/definitions/dao.ListAccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "创建访问令牌",
                "parameters": [
                    {
                        "description": "创建访问令牌请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.CreateAccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dao.CreateAccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/access-token/{token_id}": {
            "get": {
                "description": "获取访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "获取访问令牌",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "访问令牌ID",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dao.AccessTokenSpec"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "删除访问令牌",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "访问令牌ID",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/agent/jobs": {
            "get": {
                "description": "获取设备的作业列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "获取设备的作业列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dao.ListJobsResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/agent/register": {
            "post": {
                "description": "注册设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "注册设备",
                "parameters": [
                    {
                        "description": "注册请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/dao.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "冲突",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/agent/unregister": {
            "post": {
                "description": "注销设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "注销设备",
                "responses": {
                    "200": {
                        "description": "注销成功"
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/device": {
            "get": {
                "description": "列出设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "列出设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页起始位置",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页每页数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "列出成功",
                        "schema": {
                            "$ref": "#/definitions/dao.ListDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/device/{device_id}": {
            "get": {
                "description": "获取设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "获取设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dao.DeviceSpec"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备"
                ],
                "summary": "删除设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dao.AccessTokenSpec": {
            "type": "object",
            "required": [
                "accessToken"
            ],
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "deviceUuid": {
                    "type": "string"
                },
                "expireTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "dao.CreateAccessTokenRequest": {
            "type": "object",
            "properties": {
                "expireTime": {
                    "type": "string"
                }
            }
        },
        "dao.CreateAccessTokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "dao.DetectOptions": {
            "type": "object",
            "required": [
                "modelName"
            ],
            "properties": {
                "confThreshold": {
                    "type": "number"
                },
                "interval": {
                    "type": "integer"
                },
                "iouThreshold": {
                    "type": "number"
                },
                "labels": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string"
                }
            }
        },
...
        "dao.DeviceSpec": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lastPingTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "registerTime": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dao.JobSpec": {
            "type": "object",
            "required": [
                "createTime",
                "input",
                "kind",
                "status",
                "updateTime",
                "uuid"
            ],
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "detect": {
                    "$ref": "#/definitions/dao.DetectOptions"
                },
                "device": {
                    "$ref": "#/definitions/dao.DeviceSpec"
                },
                "id": {
                    "type": "integer"
                },
                "input": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/model.JobKind"
                },
                "query": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                },
                "updateTime": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "videoSegment": {
                    "$ref": "#/definitions/dao.VideoSegmentOptions"
                },
                "workflowId": {
                    "type": "integer"
                }
            }
        },
        "dao.ListAccessTokenResponse": {
            "type": "object",
            "properties": {
                "accessTokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.AccessTokenSpec"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dao.ListDeviceResponse": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.DeviceSpec"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dao.ListJobsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.JobSpec"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dao.RegisterRequest": {
            "type": "object",
            "required": [
                "accessToken",
                "name"
            ],
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dao.RegisterResponse": {
            "type": "object",
            "properties": {
                "s3AccessKeyID": {
                    "type": "string"
                },
                "s3SecretAccessKey": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "dao.VideoSegmentOptions": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "integer"
                }
            }
        },
...
        "model.JobKind": {
            "type": "string",
            "enum": [
                "detect",
                "video_segment"
            ],
            "x-enum-varnames": [
                "JobKindDetect",
                "JobKindVideoSegment"
            ]
        },
        "model.JobStatus": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "JobStatusStopped",
                "JobStatusRunning"
            ]
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        }
    }
}